Force logout: check if user exists

+++ force-logout.sh
@@ -3,5 +3,9 @@
 if [[ -z $1 ]]; then
     echo "No user specified!"
 else
-    sudo pkill -KILL -u "$1"
+    if id "$1" &>/dev/null; then
+        sudo pkill -KILL -u "$1"
+    else
+        echo "User $1 does not exist!"
+    fi
 fi
